requirements.txt
Flask
web3
flask_cors
Werkzeug
fpdf
openpyxl
psycopg2-binary==2.9.7
python-dotenv==1.0.0
pyjwt==2.8.0
gunicorn==21.2.0